Evolução Cadastro do Perfil Usuário

+++ src/main/java/school/sptech/zup/dto/response/AnalisePerfilResponse.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class AnalisePerfilResponse {
     private Integer idUsuario;
-    private Integer idPerfil;
+    private Long idPerfil;
     private String Perfil;
 }
